# specs/001-petiatrics-platform-all/contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: Petiatrics API
  version: 1.0.0
  description: >-
    Phase 0 API contract for the Petiatrics multi-tenant clinic platform.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
...
paths:
...
  /patients/{patientId}/visits:
    post:
      tags: [Visits]
      summary: Create a visit record draft.
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVisitRequest'
      responses:
        '201':
          description: Visit draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeVisitRecord'
  /patients/{patientId}/visits/{visitId}:
    patch:
      tags: [Visits]
      summary: Update a visit record.
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/VisitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVisitRequest'
      responses:
        '200':
          description: Visit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeVisitRecord'
  /patients/{patientId}/visits/{visitId}/finalize:
    post:
      tags: [Visits]
      summary: Finalize a visit record and make it billing-eligible.
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/VisitId'
      responses:
        '200':
          description: Visit finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeVisitRecord'
...
  /audit/logs:
    get:
      tags: [Audit]
      summary: Query audit log entries.
      parameters:
        - in: query
          name: entityType
          schema:
            type: string
        - in: query
          name: actorId
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Audit log list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeAuditLogList'
components:
  parameters:
...
    PatientId:
      in: path
      name: patientId
      required: true
      schema:
        type: string
    VisitId:
      in: path
      name: visitId
      required: true
      schema:
        type: string
...
  schemas:
...
    EnvelopeVisitRecord:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/VisitRecord'
        meta:
          type: object
        error:
          type: ['null', object]
...
    EnvelopeAuditLogList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        error:
          type: ['null', object]
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    CreateVisitRequest:
      type: object
      required: [soap]
      properties:
        soap:
          $ref: '#/components/schemas/SOAP'
        prescriptions:
          type: array
          items:
            $ref: '#/components/schemas/Prescription'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
    UpdateVisitRequest:
      type: object
      properties:
        soap:
          $ref: '#/components/schemas/SOAP'
        prescriptions:
          type: array
          items:
            $ref: '#/components/schemas/Prescription'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        amendmentReason:
          type: string
...
    VisitRecord:
      type: object
      properties:
        id:
          type: string
        patientId:
          type: string
        vetId:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, finalized, amended]
        soap:
          $ref: '#/components/schemas/SOAP'
        prescriptions:
          type: array
          items:
            $ref: '#/components/schemas/Prescription'
    SOAP:
      type: object
      properties:
        subjective:
          type: string
        objective:
          type: string
        assessment:
          type: string
        plan:
          type: string
    Prescription:
      type: object
      properties:
        drug:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        productId:
          type: string
          format: uuid
          nullable: true
        inventoryLinked:
          type: boolean
    Attachment:
      type: object
      properties:
        type:
          type: string
          enum: [lab_result, imaging, file]
        url:
          type: string
          format: uri
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        operation:
          type: string
        actorId:
          type: string
          format: uuid
        actorRole:
          type: string
        timestamp:
          type: string
          format: date-time
...